requirements.txt
jupyterlab
numpy~=1.24.2
matplotlib~=3.7.1
torch~=2.0.0
torchvision
torchsummary
torchmetrics~=0.11.4
lightning~=2.0.0
sklearn
scipy~=1.10.1
tensorboard
ipywidgets
seaborn~=0.12.2
einops
Pillow~=9.4.0
pandas~=2.0.3